Changed githooks, github actions to run tests with all features
