+++ .githooks/pre-commit
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-cargo check --all-targets --features testing
+cargo check --all-targets --all-features
