+++ .github/workflows/ci.yml
@@ -84,29 +84,16 @@ jobs:
         name: Prepare LD_LIBRARY_PATH (Ubuntu only)
         run: echo LD_LIBRARY_PATH=${pythonLocation}/lib >> $GITHUB_ENV
 
-      - name: Build docs
-        run: cargo doc --no-default-features --features testing --verbose --target ${{ matrix.platform.rust-target }}
-
       - name: Build (no features)
         run: cargo build --no-default-features --verbose --target ${{ matrix.platform.rust-target }}
 
       - name: Build (all additive features)
-        run: cargo build --no-default-features --features testing --verbose --target ${{ matrix.platform.rust-target }}
+        run: cargo build --all-features --verbose --target ${{ matrix.platform.rust-target }}
 
       # Run tests (except on PyPy, because no embedding API).
       - if: matrix.python-version != 'pypy-3.6'
         name: Test
-        run: cargo test --no-default-features --features testing --target ${{ matrix.platform.rust-target }}
-
-      # Run tests again, but in abi3 mode
-      - if: matrix.python-version != 'pypy-3.6'
-        name: Test (abi3)
-        run: cargo test --no-default-features --features testing --target ${{ matrix.platform.rust-target }}
-
-      # Run tests again, for abi3-py36 (the minimal Python version)
-      - if: (matrix.python-version != 'pypy-3.6') && (matrix.python-version != '3.6')
-        name: Test (abi3-py36)
-        run: cargo test --no-default-features --features testing --target ${{ matrix.platform.rust-target }}
+        run: cargo test --all-features --target ${{ matrix.platform.rust-target }}
 
       - name: Install python test dependencies
         run: |
@@ -138,7 +125,7 @@ jobs:
       - uses: actions-rs/cargo@v1
         with:
           command: test 
-          args: --features testing
+          args: --all-features
         env:
           CARGO_INCREMENTAL: 0
           RUSTFLAGS: "-Zprofile -Ccodegen-units=1 -Cinline-threshold=0 -Clink-dead-code -Coverflow-checks=off -Cpanic=abort -Zpanic_abort_tests"
