+++ pytests/test_tokio_asyncio.rs
@@ -1,3 +1,5 @@
+mod common;
+
 use std::{
     future::{pending, Future},
     thread,
@@ -32,16 +34,6 @@ fn sleep_for(py: Python, secs: &PyAny) -> PyResult<PyObject> {
     })
 }
 
-const TEST_MOD: &'static str = r#"
-import asyncio 
-
-async def py_sleep(duration):
-    await asyncio.sleep(duration)
-
-async def sleep_for_1s(sleep_for):
-    await sleep_for(1)
-"#;
-
 fn test_into_coroutine(
     py: Python,
 ) -> PyResult<impl Future<Output = PyResult<()>> + Send + 'static> {
@@ -51,8 +43,13 @@ fn test_into_coroutine(
         .as_ref(py)
         .add_wrapped(wrap_pyfunction!(sleep_for))?;
 
-    let test_mod: PyObject =
-        PyModule::from_code(py, TEST_MOD, "test_rust_coroutine/test_mod.py", "test_mod")?.into();
+    let test_mod: PyObject = PyModule::from_code(
+        py,
+        common::TEST_MOD,
+        "test_rust_coroutine/test_mod.py",
+        "test_mod",
+    )?
+    .into();
 
     Ok(async move {
         Python::with_gil(|py| {
@@ -60,24 +57,6 @@ fn test_into_coroutine(
                 py,
                 test_mod
                     .call_method1(py, "sleep_for_1s", (sleeper_mod.getattr(py, "sleep_for")?,))?
-                    .as_ref(py),
-            )
-        })?
-        .await?;
-        Ok(())
-    })
-}
-
-fn test_into_future(py: Python) -> PyResult<impl Future<Output = PyResult<()>> + Send + 'static> {
-    let test_mod: PyObject =
-        PyModule::from_code(py, TEST_MOD, "test_rust_coroutine/test_mod.py", "test_mod")?.into();
-
-    Ok(async move {
-        Python::with_gil(|py| {
-            pyo3_asyncio::into_future(
-                py,
-                test_mod
-                    .call_method1(py, "py_sleep", (1.into_py(py),))?
                     .as_ref(py),
             )
         })?
@@ -103,10 +82,6 @@ fn test_async_sleep<'p>(
     })
 }
 
-fn test_blocking_sleep() {
-    thread::sleep(Duration::from_secs(1));
-}
-
 fn main() {
     thread::spawn(|| {
         CURRENT_THREAD_RUNTIME.block_on(pending::<()>());
@@ -127,7 +102,7 @@ fn main() {
                 }),
             ),
             new_sync_test("test_blocking_sleep".into(), || {
-                test_blocking_sleep();
+                common::test_blocking_sleep();
                 Ok(())
             }),
             Test::new_async(
@@ -143,7 +118,7 @@ fn main() {
             Test::new_async(
                 "test_into_future".into(),
                 Python::with_gil(|py| {
-                    test_into_future(py)
+                    common::test_into_future(py)
                         .map_err(|e| {
                             e.print_and_set_sys_last_vars(py);
                         })
